+++ src/convert.rs
@@ -1,4 +1,0 @@
-pub trait Convert {
-    fn to_imperial(&self);
-    fn to_metric(&self);
-}