+++ src/error.rs
@@ -6,14 +6,14 @@ use {
 
 #[derive(Debug)]
 pub enum Error {
-    IoError(io::Error),
+    Io(io::Error),
     TomlDecode(de::Error),
     TomlEncode(ser::Error),
 }
 
 impl From<io::Error> for Error {
     fn from(e: io::Error) -> Self {
-        Self::IoError(e)
+        Self::Io(e)
     }
 }
 
@@ -32,7 +32,7 @@ impl From<ser::Error> for Error {
 impl fmt::Display for Error {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::IoError(e) => write!(f, "{e}"),
+            Self::Io(e) => write!(f, "{e}"),
             Self::TomlDecode(e) => write!(f, "{e}"),
             Self::TomlEncode(e) => write!(f, "{e}"),
         }
@@ -42,7 +42,7 @@ impl fmt::Display for Error {
 impl error::Error for Error {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         match self {
-            Self::IoError(e) => Some(e),
+            Self::Io(e) => Some(e),
             Self::TomlDecode(e) => Some(e),
             Self::TomlEncode(e) => Some(e),
         }