+++ src/gui/mod.rs
@@ -1,7 +1,8 @@
 #![warn(clippy::all, clippy::pedantic)]
+mod actions;
 mod adjustments;
-mod dialogs;
-pub mod file;
+pub(crate) mod dialogs;
+pub(crate) mod file;
 
 use {
     crate::{
@@ -9,13 +10,14 @@ use {
         template::Template,
         Convert, CONFIG, FILE,
     },
+    actions::Actions,
     adjustments::Adjustments,
     dialogs::Dialogs,
     file::File,
     fretboard_layout::{Handedness, Specs, Units, Variant},
     gtk::{
         gdk_pixbuf::Pixbuf,
-        gio::{Cancellable, MemoryInputStream, SimpleAction},
+        gio::{Cancellable, MemoryInputStream},
         glib,
         glib::{char::Char, clone, MainContext, OptionArg, OptionFlags, PRIORITY_DEFAULT},
         prelude::*,
@@ -24,117 +26,23 @@ use {
     std::{mem, path::PathBuf, process::Command, rc::Rc, string::ToString, thread},
 };
 
-struct Gui {
+/// The elements making up the graphical user interface
+pub struct Gui {
     window: gtk::ApplicationWindow,
     image_preview: gtk::Picture,
-    scale: gtk::Scale,
-    scale_fine: gtk::SpinButton,
+    pub(crate) scale: gtk::Scale,
+    pub(crate) scale_fine: gtk::SpinButton,
     variant: gtk::ComboBox,
     handedness: gtk::ComboBox,
     scale_multi_course: gtk::Scale,
-    scale_multi_fine: gtk::SpinButton,
+    pub(crate) scale_multi_fine: gtk::SpinButton,
     fret_count: gtk::SpinButton,
     pfret_label: gtk::Label,
     perpendicular_fret: gtk::SpinButton,
-    nut_width: gtk::SpinButton,
-    bridge_spacing: gtk::SpinButton,
+    pub(crate) nut_width: gtk::SpinButton,
+    pub(crate) bridge_spacing: gtk::SpinButton,
     dialogs: Dialogs,
-    adjustments: Adjustments,
-}
-
-struct Actions {
-    open_template: SimpleAction,
-    save: SimpleAction,
-    save_as: SimpleAction,
-    open_external: SimpleAction,
-    preferences: SimpleAction,
-    about: SimpleAction,
-    quit: SimpleAction,
-}
-
-impl Actions {
-    fn init() -> Actions {
-        Actions {
-            open_template: SimpleAction::new("open_template", None),
-            save: SimpleAction::new("save", None),
-            save_as: SimpleAction::new("save_as", None),
-            open_external: SimpleAction::new("open_external", None),
-            preferences: SimpleAction::new("preferences", None),
-            about: SimpleAction::new("about", None),
-            quit: SimpleAction::new("quit", None),
-        }
-    }
-
-    fn connect(&self, gui: &Rc<Gui>) {
-        self.open_template
-            .connect_activate(clone!(@weak gui => move |_, _| {
-                gui.dialogs.open_template.show();
-            }));
-
-        self.save.connect_activate(clone!(@weak gui => move |_, _| {
-            gui.save();
-        }));
-
-        self.save_as
-            .connect_activate(clone!(@weak gui => move |_, _| {
-                gui.dialogs.save_as.show();
-            }));
-
-        self.open_external
-            .connect_activate(clone!(@weak gui => move |_, _| {
-                if let Ok(file) = FILE.try_lock() {
-                    if !file.saved() {
-                        gui.dialogs.save_as.show();
-                    }
-                    Gui::open_external(&file);
-                };
-            }));
-
-        self.preferences
-            .connect_activate(clone!(@weak gui => move |_, _| {
-                gui.dialogs.preferences.show();
-            }));
-
-        self.about
-            .connect_activate(clone!(@strong gui => move |_, _| {
-                gui.dialogs.about.show();
-            }));
-
-        self.quit.connect_activate(clone!(@weak gui => move |_, _| {
-            gui.cleanup();
-            gui.window.close();
-        }));
-    }
-}
-
-impl Convert for Gui {
-    fn to_metric(&self) {
-        self.adjustments.to_metric();
-        self.bridge_spacing
-            .set_value(self.bridge_spacing.value() * 20.4);
-        self.nut_width.set_value(self.nut_width.value() * 20.4);
-        self.scale.set_value(self.scale.value() * 20.4);
-        self.scale_multi_fine
-            .set_value(self.scale_multi_fine.value() * 20.4);
-        self.bridge_spacing.set_digits(2);
-        self.nut_width.set_digits(2);
-        self.scale_fine.set_digits(2);
-        self.scale_multi_fine.set_digits(2);
-    }
-
-    fn to_imperial(&self) {
-        self.adjustments.to_imperial();
-        self.bridge_spacing
-            .set_value(self.bridge_spacing.value() / 20.4);
-        self.nut_width.set_value(self.nut_width.value() / 20.4);
-        self.scale.set_value(self.scale.value() / 20.4);
-        self.scale_multi_fine
-            .set_value(self.scale_multi_fine.value() / 20.4);
-        self.bridge_spacing.set_digits(3);
-        self.nut_width.set_digits(3);
-        self.scale_fine.set_digits(3);
-        self.scale_multi_fine.set_digits(3);
-    }
+    pub(crate) adjustments: Adjustments,
 }
 
 impl Gui {
@@ -161,25 +69,24 @@ impl Gui {
         }
     }
 
-    fn add_actions(&self, app: &gtk::Application) -> Actions {
-        let actions = Actions::init();
-
-        app.set_accels_for_action("win.open_template", &["<primary>O"]);
-        app.set_accels_for_action("win.save", &["<primary>S"]);
-        app.set_accels_for_action("win.save_as", &["<primary><shift>S"]);
-        app.set_accels_for_action("win.open_external", &["<primary>E"]);
-        app.set_accels_for_action("win.preferences", &["<primary><shift>P"]);
-        app.set_accels_for_action("win.about", &["<primary>A"]);
-        app.set_accels_for_action("win.quit", &["<primary>Q"]);
-
-        self.window.add_action(&actions.open_template);
-        self.window.add_action(&actions.save);
-        self.window.add_action(&actions.save_as);
-        self.window.add_action(&actions.open_external);
-        self.window.add_action(&actions.preferences);
-        self.window.add_action(&actions.about);
-        self.window.add_action(&actions.quit);
-        actions
+    fn setup(&self) {
+        let cfg = CONFIG.try_lock().unwrap().clone();
+        let units = cfg.units;
+        if units == Units::Imperial {
+            self.adjustments.to_imperial();
+        }
+        let mut statefile = config::get_config_dir();
+        statefile.push("state.toml");
+        match Template::load_from_file(statefile) {
+            Ok(t) => self.load_template(&t),
+            Err(e) => eprintln!("Error loading statefile: {e}"),
+        }
+        self.toggle_multi();
+        self.draw_preview(false);
+        self.window.set_title(Some(&format!(
+            "Gfret - {} - <unsaved>",
+            env!("CARGO_PKG_VERSION")
+        )));
     }
 
     fn get_handedness(&self) -> Handedness {
@@ -280,7 +187,7 @@ impl Gui {
     }
 
     /// Sets widget state to match temmplate
-    pub fn load_template(&self, template: &Template) {
+    pub(crate) fn load_template(&self, template: &Template) {
         self.scale.set_value(template.scale);
         self.fret_count.set_value(template.count.into());
         match template.scale_treble {
@@ -441,6 +348,7 @@ impl Gui {
     }
 }
 
+/// Runs the graphical interface
 pub fn run(template: Option<&str>) {
     let template = template.map(ToString::to_string);
     let application = gtk::Application::new(
@@ -469,30 +377,9 @@ pub fn run(template: Option<&str>) {
 
 fn build_ui(application: &Application) -> Rc<Gui> {
     let gui = Rc::new(Gui::init());
-    let cfg = CONFIG.try_lock().unwrap().clone();
-    let units = cfg.units;
-    if units == Units::Imperial {
-        gui.adjustments.to_imperial();
-    }
-    let mut statefile = config::get_config_dir();
-    statefile.push("state.toml");
-    if statefile.exists() {
-        match Template::load_from_file(statefile) {
-            Ok(t) => gui.load_template(&t),
-            Err(e) => eprintln!("Error loading statefile: {e}"),
-        }
-    }
-
-    gui.add_actions(application).connect(&gui);
-
-    gui.window.set_title(Some(&format!(
-        "Gfret - {} - <unsaved>",
-        env!("CARGO_PKG_VERSION")
-    )));
-
     gui.window.set_application(Some(application));
-    gui.toggle_multi();
-    gui.draw_preview(false);
+    gui.setup();
+    Actions::default().add(&gui, application);
 
     gui.scale
         .connect_value_changed(clone!(@weak gui => move |_| {