+++ src/lib.rs
@@ -1,0 +1,103 @@
+#![warn(clippy::all, clippy::pedantic)]
+// switch to this when it stabilizes
+//#![feature(mutex_unlock)]
+#![doc = include_str!("../README.md")]
+
+use {fretboard_layout::Config, gtk::prelude::*, gui::file::File, std::sync::Mutex};
+/// The cli
+pub mod cli;
+/// Handles getting the configuration data to and from disk
+mod config;
+/// Crate specific errors
+pub(crate) mod error;
+/// The Gtk user interface to gfret.
+pub mod gui;
+/// Persistent templates
+mod template;
+
+pub use gui::{dialogs::PrefWidgets, Gui};
+
+#[macro_use]
+extern crate lazy_static;
+
+lazy_static! {
+    static ref CONFIG: Mutex<Config> =
+        Mutex::new(config::Config::from_file().unwrap_or_default().truncate());
+    static ref FILE: Mutex<File> = Mutex::new(File::default());
+}
+
+/// Switches between imperial and metric units
+pub(crate) trait Convert {
+    /// Changes to imperial units
+    fn to_imperial(&self);
+    /// Changes to metric units
+    fn to_metric(&self);
+}
+
+impl Convert for Gui {
+    fn to_metric(&self) {
+        self.adjustments.to_metric();
+        self.bridge_spacing
+            .set_value(self.bridge_spacing.value() * 20.4);
+        self.nut_width.set_value(self.nut_width.value() * 20.4);
+        self.scale.set_value(self.scale.value() * 20.4);
+        self.scale_multi_fine
+            .set_value(self.scale_multi_fine.value() * 20.4);
+        self.bridge_spacing.set_digits(2);
+        self.nut_width.set_digits(2);
+        self.scale_fine.set_digits(2);
+        self.scale_multi_fine.set_digits(2);
+    }
+
+    fn to_imperial(&self) {
+        self.adjustments.to_imperial();
+        self.bridge_spacing
+            .set_value(self.bridge_spacing.value() / 20.4);
+        self.nut_width.set_value(self.nut_width.value() / 20.4);
+        self.scale.set_value(self.scale.value() / 20.4);
+        self.scale_multi_fine
+            .set_value(self.scale_multi_fine.value() / 20.4);
+        self.bridge_spacing.set_digits(3);
+        self.nut_width.set_digits(3);
+        self.scale_fine.set_digits(3);
+        self.scale_multi_fine.set_digits(3);
+    }
+}
+
+impl Convert for PrefWidgets {
+    fn to_metric(&self) {
+        let mut val = self.border.value();
+        let mut adjustment = self.border.adjustment();
+        adjustment.set_upper(40.0);
+        adjustment.set_step_increment(0.10);
+        adjustment.set_page_increment(5.0);
+        self.border.set_value(val * 20.4);
+        self.border.set_digits(2);
+
+        val = self.line_weight.value();
+        adjustment = self.line_weight.adjustment();
+        adjustment.set_upper(2.0);
+        adjustment.set_step_increment(0.10);
+        adjustment.set_page_increment(0.50);
+        self.line_weight.set_value(val * 20.4);
+        self.line_weight.set_digits(2);
+    }
+
+    fn to_imperial(&self) {
+        let mut val = self.border.value();
+        let mut adjustment = self.border.adjustment();
+        adjustment.set_upper(40.0 / 20.4);
+        adjustment.set_step_increment(0.01);
+        adjustment.set_page_increment(0.10);
+        self.border.set_value(val / 20.4);
+        self.border.set_digits(3);
+
+        val = self.line_weight.value();
+        adjustment = self.line_weight.adjustment();
+        adjustment.set_upper(0.098);
+        adjustment.set_step_increment(0.01);
+        adjustment.set_page_increment(0.05);
+        self.line_weight.set_value(val / 20.4);
+        self.line_weight.set_digits(3);
+    }
+}